moves ballot component into it own folder and resolves related imports
